Move cifV2ToV1 from cif.py to cryspy/parser.py
