+++ easydiffraction/calculators/cryspy/calculator.py
@@ -15,8 +15,8 @@ import numpy as np
 from cryspy.procedure_rhochi.rhochi_by_dictionary import rhochi_calc_chi_sq_by_dictionary
 from easyscience import global_object as borg
 
-from easydiffraction.io.cif import cifV2ToV1
-from easydiffraction.io.cif import cifV2ToV1_tof
+from easydiffraction.calculators.cryspy.parser import cifV2ToV1
+from easydiffraction.calculators.cryspy.parser import cifV2ToV1_tof
 from easydiffraction.calculators.cryspy.parser import CryspyParser
 
 # from pathos import multiprocessing as mp
